+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Fixes:**
 
+- Fixed issue when remembered comment was for all pages/posts, not per specific one, #462
 - Now sorting is treated properly on real-time update feature, #460
 - Resolved issue with timezone when was writing "n hours ago" on a new comment, #464